requirements.txt
fal-client>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.1
